requirements.txt
fastapi
# "uvicorn[standard]"
uvicorn
openai
python-dotenv